Fix for Bug#22988922, GETLENGTH() RETURNS -1 FOR LONGBLOB AND LONGTEXT FIELDS.

+++ src/main/java/com/mysql/cj/api/x/Warning.java
@@ -35,7 +35,7 @@ public interface Warning {
     /**
      * Warning code.
      */
-    int getCode();
+    long getCode();
 
     /**
      * Warning message.
